VOL-156: Factored out the getwithall method on the Project API.
Commit d133cb536406bd54de34b1df8130ca04dda7ba7f introduced the ability to pull back basic profile information
along with the details of a volunteer project. This seemed useful enough that it should be part of the regular
"get" API, and I couldn't think of a valid reason that this needed to be a separate API.

The original implementation checks permissions (can we update the requested project?), but I posit that it's
counter-intuitive to check for update permissions in a get method. I'm not sure I see a real need for such a
check, but if needed a similar result could be achieved by fleshing out the "context" concept in the "get"
API.