Proof-of-concept auto-updating documentation section.
Documentation pages are named with a three-number prefix to control the order
in which the table of contents page displays them. The number should be
incremented by 10 if possible to allow for reordering later without having
to rename every file. Additional categories should be added to _config.yml
as needed.

The section of the documentation which outlines how to contribute to
CiviVolunteer's documentation should include some version of the instructions
above.