VOL-206: Cleaned up CSS class names and logic around cart floating.
I imagine the removed classes (floating_cart_fixed and floating_cart_custom) were placed as future extension
points, but I'd rather put them there when we need them than introduce YAGNI. Without a real use case or
developer documentation, they run the risk of confusing matters and becoming cruft.

In the JavaScript logic, I found the use of "fixed" in variable names to be confusing. It could mean that
the cart is fixed (to the page) versus floating, or it could mean that it is fixed to the viewport (as in
the CSS position). It did in fact appear to be used both ways.