VOL-263: Removed support from api.VolunteerProjectContact.get for reporting whether or not the current user can read the contacts in the result.
This "feature" is a vestige of an old approach to determining whether or not to display the relevant form elements on the
Angular form used to create and edit Projects. The old approach was based on displaying the fields if the user had access to the
contacts in question and on managing default values on the client side. It was prone to problems such as requiring fields that were not
displayed to the user. The new approach uses discrete, explicit permissions instead.

Note: This is technically an API breakage. I think this is okay since this particular API is not intended for public consumption but
rather is necessitated by our client/server architecture.