Fix PHP 7.1 Error: Cannot pass parameter 1 by reference
Before this fix, visiting the "Manage events" page (when at least one
event has an associated volunteer project, and the server is running
PHP 7.1) would give a fatal error "Cannot pass parameter 1 by
reference".

This fix changes the call to copyValues so that it passes a variable,
making this code compatible with PHP 7.1 and above.

Here is a comparison of this behavior in different PHP versions:
https://3v4l.org/b9Jea