Regenerate DAO files using civix (#548)
* Regenerate DAO files using civix

CiviVolunteer originally used a hacked version of setup.sh to generate DAO files and SQL from the xml.
That's no longer necessary now, with the command: civix generate:entity-boilerplate
Just had to reorganize the files a little.

* Switch installer to use generated auto_install.sql and update volunteer_need.created column

Switched to the auto-install and auto-uninstall sql files, which required fixing the default value
of the created column. Also updated the BAO and added an upgrade step to fix the column.

This required a core version compatibility bump to 5.28 which requires the mysql version which
supports multiple CURRENT_TIMESTAMP defaults in a single table.